main/requirements.txt
celery
cryptography
Django
pyroute2
pyTelegramBotAPI
python-dateutil
redis
gunicorn
qrcode
Pillow
djangorestframework
